requirements.txt
elasticsearch==7.17.0
Flask==2.3.2
langchain==0.0.239
pymilvus==2.2.13
Requests==2.31.0
loguru==0.7.0
PyYAML==6.0
pypdf==3.15.5
docx2txt==0.8
unstructured==0.10.12
selenium==4.12.0
gradio==3.23.0
cohere==4.34